+++ fish2eod/tests/test_complex_condition.py
@@ -30,6 +30,32 @@ def central_source_model():
     return CentralCircularSource
 
 
+@pytest.fixture(scope="session")
+def complex_geometry_model():
+    class ComplexGeometryModel(QESModel):
+        def create_geometry(self, **kwargs) -> None:
+            background = Rectangle([-0.5, -0.5], 1, 1)
+            ground = Circle([-0.4, 0.2], 0.02)
+            bottom_source = Circle([0, -0.2], 0.1)
+            top_source = Circle([0.2, 0.3], 0.05)
+
+            self.model_geometry.add_domain("bkg", background, sigma=1)
+            self.model_geometry.add_domain("ground", ground, sigma=2)
+            self.model_geometry.add_domain("bottom_source", bottom_source, sigma=3)
+            self.model_geometry.add_domain("top_source", top_source, sigma=4)
+
+        def get_neumann_conditions(self, **kwargs):
+            return [
+                BoundaryCondition(1, self.model_geometry["bottom_source"]),
+                BoundaryCondition(-2, self.model_geometry["top_source"]),
+            ]
+
+        def get_dirichlet_conditions(self, **kwargs):
+            return [BoundaryCondition(0, self.model_geometry["ground"])]
+
+    return ComplexGeometryModel
+
+
 @pytest.mark.integration
 @pytest.mark.parametrize(
     "bkg_cond, source_cond, source, name",
